Move files to target folder
