--- lab_3/decrypted_text.txt
@@ -1,1 +1,0 @@
-I walk a lonely road The only one that I have ever known Don't know where it goes But it's home to me, and I walk alone I walk this empty street On the Boulevard of Broken Dreams Where the city sleeps And I'm the only one, and I walk alone My shadow's the only one that walks beside me My shallow heart's the only thing that's beating Sometimes, I wish someone out there will find me 'Til then, I walk alone I'm walking down the line That divides me somewhere in my mind On the borderline Of the edge, and where I walk alone Read between the lines What's fucked up, and everything's alright Check my vital signs To know I'm still alive, and I walk alone My shadow's the only one that walks beside me My shallow heart's the only thing that's beating Sometimes, I wish someone out there will find me 'Til then, I walk alone
