--- lab_3/private.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEAwi0cuavrDGrN+QcYN+FGw1L6xa7ScmQ/9c5RSZazVoqisEZZ
-7hDwL2dWHhl9GzrM1JJLyPdBiXaXJ5DAQd6ceurBGXQn1ntYxugL4mfXViCAy5gv
-cZiP36xn5aozy2HjJKcK2Z0DP+S8q4HeJU09Th6DS4lQmWnvXiTpxinHMghYJVeK
-xkJ0ADihDQCODk8z0nOW7xGeXoyRoohOdxLIuPF6Z2fUhN32EnFCyM+jyHNGj87u
-nluBNSqggFCILKfQCmSt0kfs0fcR1pyMJSmg/9weK4Ygxq6RR20apFntGvQs23EB
-AjLKkGoqLZxToyF3cqwTKKOqSox25oiKyaU7UQIDAQABAoIBAAM3DkFUQ4RtfcrU
-Z14gfqSUCocsp3shpg32vDGS21YQ3lP7HDjVArcUN9CwhRnm8baI1BSoJGeHZUYQ
-T2Cf1Bk0ax1V5yV13iDP/YdihwYb7GoCk3GG8EXos8ExneEtlh5YnjU/dGrxyX5w
-/GnDc0MIfg3JgKJDpuUZCR5g+XV+drUN9+NzU9hxbVJsTyoLPeaX9vlxM0iDS8HC
-9AJLj5eSYDDiqHy7eOu0zmOOxAWmkTeijq0QEbgbrc42cwqxz2EQVOJShey32XoH
-kwX6XE4vYYCvn001FeIGdkLfu0/hW5hbH21/txmo8Q0pAuZmiHmcJ7KNpC1zFPPf
-sf0BEEkCgYEA9cxy6C5eOEcAtBcF7ioba0Gox4t3ECog9o+NCdngcCjxI1Nsgq/P
-PjZKQzlZTpw282fhAN0Bq7wxWPElvmWLKX8Ek9BIYIP6S8ODud9g9AFLcauLRQZp
-DLp1n1lNdoAVnXeTsE2MpeHl0Wp60ULHN5Et5VMm8n+Swi/ZDSxhN8kCgYEAyjww
-Db+fRoqXBr2fa/6lr4nGotKXWM7ALXKJnCPM9R8XyAzpCV9YHIZuZl00bxWgHp3a
-VUtzxvsTqedH+1/97zURfo+SFUG+U9ws1RddcYC/ZnI1T1T0qk0CvPJJUM1f6F55
-uHsNKClU6C0SksOSVrsns6Nzo79443cZr6QJO0kCgYEAqgud9wiVAEQGtZAz9xdE
-Tshm1l2iqhkrIsge35DOl2mLoU/GgwbTzgtLBwS2xw6FaZEU9T7wN/oC/bAe14By
-bOvFo+j1GWhUkXOaPs3NyiDXK7VsrP0JUdfxJaNnVrK3wW+HSfHsepo0HtOWUgzr
-u4KBiOQkzqD5S8REWbu5KhkCgYBFHbjWmgzXUPo4pYncoNUDXzTx5MMw6MgpQCHj
-qSCdDZlVF4R1XY6BgTgt5+DJrjxg+lXqHLVe5eRchtzE/pOMszhdZrfoZ+WiR3dT
-BH89EWFYo+OwTzqnEVbEOJHx+Z6zjcZP1E400ZthFjZozkovNhtle5FD/2ZTZxry
-Oj+lcQKBgQCvo0IbeOil9gGRlrlpaF3a7J54ZPBDw3BD1H+FSqu85QIdWNnSkV+/
-N/sRah/78CvrX3mCHkgQNc1EfRAUMD3EkYF8LW5bB7eJ4wKcBL2pA/A2FcjFc2G7
-vajzKbjRgeHTdWPXioGu+A/dwo2PmttmynS9tXUNISuRB4eJ/rgIrQ==
------END RSA PRIVATE KEY-----
