--- lab_3/public.pem
@@ -1,9 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAwi0cuavrDGrN+QcYN+FG
-w1L6xa7ScmQ/9c5RSZazVoqisEZZ7hDwL2dWHhl9GzrM1JJLyPdBiXaXJ5DAQd6c
-eurBGXQn1ntYxugL4mfXViCAy5gvcZiP36xn5aozy2HjJKcK2Z0DP+S8q4HeJU09
-Th6DS4lQmWnvXiTpxinHMghYJVeKxkJ0ADihDQCODk8z0nOW7xGeXoyRoohOdxLI
-uPF6Z2fUhN32EnFCyM+jyHNGj87unluBNSqggFCILKfQCmSt0kfs0fcR1pyMJSmg
-/9weK4Ygxq6RR20apFntGvQs23EBAjLKkGoqLZxToyF3cqwTKKOqSox25oiKyaU7
-UQIDAQAB
------END PUBLIC KEY-----
