+++ lab_6/asymmetrical.py
