+++ lab_6/main.py
