+++ lab_6/symmetrical.py
