+++ lab_6/work_file.py
